refactor: move alias handling opts to options
