+++ src/ape/options.py
@@ -1,6 +1,8 @@
 import click
+from typing import List
 
-from ape import networks
+from ape import accounts, networks
+from ape.exceptions import AliasAlreadyInUseError
 
 
 class NetworkChoice(click.Choice):
@@ -31,3 +33,30 @@ def verbose_option(help=""):
         default=False,
         help=help,
     )
+
+
+class Alias(click.Choice):
+    """Wraps ``click.Choice`` to load account aliases for the active project at runtime."""
+
+    name = "alias"
+
+    def __init__(self):
+        # NOTE: we purposely skip the constructor of `Choice`
+        self.case_sensitive = False
+
+    @property
+    def choices(self) -> List[str]:  # type: ignore
+        # NOTE: This is a hack to lazy-load the aliases so CLI invocation works properly
+        return list(accounts.aliases)
+
+
+def _require_non_existing_alias(arg):
+    if arg in accounts.aliases:
+        raise AliasAlreadyInUseError(arg)
+    return arg
+
+
+existing_alias_argument = click.argument("alias", type=Alias())
+non_existing_alias_argument = click.argument(
+    "alias", callback=lambda ctx, param, arg: _require_non_existing_alias(arg)
+)
