+++ src/ape_accounts/_cli.py
@@ -1,31 +1,15 @@
 import json
-from typing import List
 
 import click
 from eth_account import Account as EthAccount  # type: ignore
 from eth_utils import to_bytes
 
 from ape import accounts
-from ape.exceptions import AliasAlreadyInUseError
+from ape.options import existing_alias_argument, non_existing_alias_argument
 from ape.utils import Abort, notify
 
 # NOTE: Must used the instantiated version of `AccountsContainer` in `accounts`
 container = accounts.containers["accounts"]
-
-
-class Alias(click.Choice):
-    """Wraps ``click.Choice`` to load account aliases for the active project at runtime."""
-
-    name = "alias"
-
-    def __init__(self):
-        # NOTE: we purposely skip the constructor of `Choice`
-        self.case_sensitive = False
-
-    @property
-    def choices(self) -> List[str]:  # type: ignore
-        # NOTE: This is a hack to lazy-load the aliases so CLI invocation works properly
-        return list(accounts.aliases)
 
 
 @click.group(short_help="Manage local accounts")
@@ -56,11 +40,8 @@ def _list():
 
 
 @cli.command(short_help="Create a new keyfile account with a random private key")
-@click.argument("alias")
+@non_existing_alias_argument
 def generate(alias):
-    if alias in accounts.aliases:
-        raise AliasAlreadyInUseError(alias)
-
     path = container.data_folder.joinpath(f"{alias}.json")
     extra_entropy = click.prompt(
         "Add extra entropy for key generation...",
@@ -79,11 +60,8 @@ def generate(alias):
 
 # Different name because `import` is a keyword
 @cli.command(name="import", short_help="Add a new keyfile account by entering a private key")
-@click.argument("alias")
+@non_existing_alias_argument
 def _import(alias):
-    if alias in accounts.aliases:
-        raise AliasAlreadyInUseError(alias)
-
     path = container.data_folder.joinpath(f"{alias}.json")
     key = click.prompt("Enter Private Key", hide_input=True)
     try:
@@ -101,7 +79,7 @@ def _import(alias):
 
 
 @cli.command(short_help="Change the password of an existing account")
-@click.argument("alias", type=Alias())
+@existing_alias_argument
 def change_password(alias):
     account = accounts.load(alias)
     account.change_password()
@@ -109,7 +87,7 @@ def change_password(alias):
 
 
 @cli.command(short_help="Delete an existing account")
-@click.argument("alias", type=Alias())
+@existing_alias_argument
 def delete(alias):
     account = accounts.load(alias)
     account.delete()
